Filter implementado em todas as urls menos a de login

+++ src/main/java/com/br/fat/controleprotocolo/controller/ControllerPessoa.java
@@ -7,6 +7,7 @@ package com.br.fat.controleprotocolo.controller;
 
 import com.br.fat.controleprotocolo.dao.PessoaDAO;
 import com.br.fat.controleprotocolo.model.Pessoa;
+import com.br.fat.controleprotocolo.model.Usuario;
 import java.util.List;
 
 /**
@@ -17,7 +18,12 @@ public class ControllerPessoa extends Controller {
 
     private final PessoaDAO pdao = new PessoaDAO();
 
+    public ControllerPessoa(Usuario u) {
+        super(u);
+    }
+
     public String createPessoa(String jsonPessoa) throws Exception {
+        this.verifyPermission(this.user.getPermissao().getWrite());
 
         Pessoa p = gson.fromJson(jsonPessoa, Pessoa.class);
 
@@ -30,20 +36,27 @@ public class ControllerPessoa extends Controller {
     }
 
     public List<Pessoa> getAllPessoa() throws Exception {
+        this.verifyPermission(this.user.getPermissao().getWrite());
+
         return pdao.selectAllPessoa();
     }
-    
-    
+
     public void removePessoa(int idPessoa) throws Exception {
+        this.verifyPermission(this.user.getPermissao().getDelete());
+
         if (idPessoa <= 0) {
             throw new Exception();
         } else {
-            pdao.deletePessoa (idPessoa);
+            pdao.deletePessoa(idPessoa);
         }
     }
 
     public void editPessoa(String jsonPessoa) throws Exception {
+        this.verifyPermission(this.user.getPermissao().getDelete());
         Pessoa p = gson.fromJson(jsonPessoa, Pessoa.class);
+        if (p.getId() <= 0) {
+            throw new Exception();
+        }
         pdao.updatePessoa(p);
     }
 }
